+++ src/main/java/com/timetrak/dto/payment/PaymentTotals.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Getter
 public class PaymentTotals {
-    private final BigDecimal totalHours;
+    private final Double totalHours;
     private final BigDecimal totalEarnings;
     private final Integer shiftsCount;
